fix error: add index.js and some errors
