+++ app.js
@@ -3,17 +3,14 @@ const express = require('express');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
+
 const { errors } = require('celebrate');
 const cors = require('cors');
-const { login, createUser } = require('./controllers/users');
-const { validationCreateUser, validationLogin } = require('./middlewares/validations');
-const auth = require('./middlewares/auth');
+
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-const NotFoundError = require('./errors/NotFoundError');
 const handleError = require('./middlewares/handleError');
-const usersRout = require('./routes/users');
-const moviesRout = require('./routes/movies');
+const limiter = require('./middlewares/limiter');
+const router = require('./routes/index');
 
 const { PORT = 3000 } = process.env;
 const app = express();
@@ -22,12 +19,6 @@ app.use(helmet());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100,
-  standardHeaders: true,
-  legacyHeaders: false,
-});
 app.use(limiter);
 app.use(requestLogger);
 app.use(cors({
@@ -36,17 +27,10 @@ app.use(cors({
     'https://helga.movies.nomoredomains.xyz',
     'http://localhost:3000'],
 }));
-app.post('/signin', validationLogin, login);
-app.post('/signup', validationCreateUser, createUser);
 
-app.use(auth);
-app.use('/users', usersRout);
-app.use('/cards', moviesRout);
+app.use(router);
 app.use(errorLogger);
 app.use(errors());
-app.use('/*', (req, res, next) => {
-  next(new NotFoundError('Запрашиваемый ресурс не найден'));
-});
 app.use(handleError);
 
 mongoose.connect('mongodb://127.0.0.1:27017/bitfilmsdb', {
