+++ middlewares/validations.js
@@ -35,6 +35,7 @@ const validationCreateUser = celebrate({
 const validationUpdateUser = celebrate({
   body: Joi.object().keys({
     name: Joi.string().min(2).max(30).required(),
+    email: Joi.string().required().email(),
   }),
 });
