+++ routes/index.js
@@ -1,0 +1,18 @@
+const router = require('express').Router();
+const auth = require('../middlewares/auth');
+const { login, createUser } = require('../controllers/users');
+const { validationCreateUser, validationLogin } = require('../middlewares/validations');
+const usersRout = require('./users');
+const moviesRout = require('./movies');
+const NotFoundError = require('../errors/NotFoundError');
+
+router.post('/signin', validationLogin, login);
+router.post('/signup', validationCreateUser, createUser);
+router.use(auth);
+router.use('/users', usersRout);
+router.use('/cards', moviesRout);
+router.use('/*', (req, res, next) => {
+  next(new NotFoundError('Запрашиваемый ресурс не найден'));
+});
+
+module.exports = router;
